Change package names for utils, reorganize. Simplified the control flow.
